Extract sync pull helpers
